+++ scopeprotocols/PeriodMeasurement.cpp
@@ -116,7 +116,7 @@ void PeriodMeasurement::Refresh()
 	float midpoint = GetAvgVoltage(din);
 
 	//Timestamps of the edges
-	vector<double> edges;
+	vector<int64_t> edges;
 	FindZeroCrossings(din, midpoint, edges);
 	if(edges.size() < 2)
 	{
@@ -127,16 +127,16 @@ void PeriodMeasurement::Refresh()
 	//Create the output
 	auto cap = new AnalogWaveform;
 
-	double rmin = FLT_MAX;
-	double rmax = 0;
+	int64_t rmin = LONG_MAX;
+	int64_t rmax = 0;
 
 	for(size_t i=0; i < (edges.size()-2); i+= 2)
 	{
 		//measure from edge to 2 edges later, since we find all zero crossings regardless of polarity
-		double start = edges[i];
-		double end = edges[i+2];
+		int64_t start = edges[i];
+		int64_t end = edges[i+2];
 
-		double delta = end - start;
+		int64_t delta = end - start;
 		cap->m_offsets.push_back(start);
 		cap->m_durations.push_back(delta);
 		cap->m_samples.push_back(delta);
